Accuracy assessment tooling (#1502)
* add notebook and function for accuracy assessments

* add olofsson methods

* complete nb example and validate

* formatting

* formatting

* formatting

* rename nb

* move supp data

* add a note on omission error limitations

* PR revisions

+++ Tools/dea_tools/classification.py
@@ -530,7 +530,7 @@ def _get_training_data_for_shp(
 
     # append ID measurement to dataset for tracking failures
     band = list(data.data_vars)[0]
-    data["_training_id"] = xr.zeros_like(data[band])
+    data["_training_id"] = xr.zeros_like(data[band], dtype=int)
     data["_training_id"] = data["_training_id"] + row["_training_id"]
 
     if "time" in data.sizes:
